backend/requirements.txt
Django>=5.0
psycopg[binary]>=3.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
python-dateutil>=2.9
gunicorn>=22.0
